Add help and settings view and modify dashboard
Adds two new views for settings and help. Cleanup some css rules and
templates. The dashboard view was refatored to show user stats and a
form to look for player to play against with. Settings are stored on
phone local storage using a new module for storage handling and help
page posses a panel which need to be fixed to work well with local
anchors. If user havent setup the app when opening the app it will
be show the settings page to do the input, navigation from dashboard
to game list page and viceversa is made also using a swipe event.
And add a new button to access settings on the application header.